everysalealbmrl/requirements.txt
fuzzywuzzy
geopandas
googlemaps
humanize
lxml
numpy
pandas
python-dotenv
pytz
requests
tweepy
